% Generated by roxygen2 (4.0.2): do not edit by hand
\name{algo3}
\alias{algo3}
\title{Generate random zones.}
\usage{
algo3(x, m, ras = FALSE)
}
\arguments{
\item{x}{an object of class \code{Raster} or a numeric vector depicting the
dimensions of a new raster.}

\item{m}{a numeric scalar indicating the number of zones to generate.}

\item{ras}{logical indicating whether output should be in raster or vector
format.  The default is \code{FALSE}.}
}
\value{
\code{algo3} returns either a \code{Raster} or
\code{SpatialPolygons} class object representing a set of randomly generated
zones.
}
\description{
\code{algo3} generates a random zoning system.  Zones can be output in either
raster or vector format.
}
\details{
Implements the algorithm outlined by Openshaw (1977).
}
\examples{
algo3(x = c(50, 50))
}
\references{
Openshaw, Stan.  1977.  "Algorithm 3: a procedure to
generate pseudo-random aggregations of N zones into M regions, where M is
less than N."  Environment and Planning A 9: 1423-1428.
}

