% Generated by roxygen2 (4.0.2): do not edit by hand
\name{algo9}
\alias{algo9}
\title{Generate autocorrelated data.}
\usage{
algo9(targI, old.r, tol = 1/ncell(old.r), maxIter = 10 * targI *
  ncell(old.r), print.out = TRUE)
}
\arguments{
\item{x}{an object of class \code{Raster} or a numeric vector depicting the
dimensions of a new raster.}

\item{mori}{a numeric scalar indicating the desired degree of
autocorrelation.  Autocorrelation is measured using Moran's I.}
}
\value{
\code{algo9} returns a \code{Raster} class object representing
a randomly generated surface with a specificed degree of autocorrelation.
}
\description{
\code{algo9} generates a valued raster with a specified degree of
autocorrelation.
}
\details{
Implements the algorithm outlined by Goodchild (1980).
}
\examples{
algo9(x = c(50, 50), mori = .3)
}
\references{
Goodchild, Michael.  1980.  "Algorithm 9: Simulation of
autocorrelation for aggregate data"  Environment and Planning A 12:
1073-1081.
}

