% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maup-package.R
\docType{package}
\name{maup-package}
\alias{maup}
\alias{maup-package}
\title{maup: Sampling Solutions to the Modifiable Areal Unit Problem}
\description{
Downloads simulated Census tract maps and plans for every U.S. county across the 2000, 2010, and 2020 decennial Censuses from the Harvard Dataverse. This data supports the study of the Modifiable Areal Unit Problem by allowing researchers to quantify how sensitive any tract-level statistic is to the choice of geographic boundaries. Builds on data from Brown, Kenny, and Simko (2025) \doi{10.31235/osf.io/nv5up_v2}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alarm-redist/maup/}
  \item Report bugs at \url{https://github.com/alarm-redist/maup/issues/}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

Authors:
\itemize{
  \item Jacob R. Brown \email{jbrown13@bu.edu} (\href{https://orcid.org/0000-0001-8037-5360}{ORCID})
  \item Tyler Simko \email{tsimko@umich.edu} (\href{https://orcid.org/0000-0001-9382-7346}{ORCID})
}

}
\keyword{internal}
