% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{maup_cache_clear}
\alias{maup_cache_clear}
\title{Clear the \code{maup} cache}
\usage{
maup_cache_clear(force = FALSE)
}
\arguments{
\item{force}{If \code{FALSE} (the default), asks for confirmation interactively.
Does not clear the cache if the session is not interactive and \code{force} is
\code{FALSE}.}
}
\value{
The path to the cache directory, invisibly.
}
\description{
Deletes all files in the cache directory.
If \code{force} is \code{FALSE} and the session is interactive, asks for confirmation
first.
}
\examples{
maup_cache_clear()
}
