% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{maup_cache_path}
\alias{maup_cache_path}
\title{Return the path to the \code{maup} cache}
\usage{
maup_cache_path()
}
\value{
A character string giving the cache directory path.
}
\description{
Returns the path to the directory where downloaded files are cached.
Use \code{options(maup.use_cache = TRUE)} to enable persistent caching, or
\code{options(maup.cache_dir = '/path/to/dir')} to specify a custom location.
}
\examples{
maup_cache_path()
}
