% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{maup_cache_size}
\alias{maup_cache_size}
\title{Report the size of the \code{maup} cache}
\usage{
maup_cache_size()
}
\value{
The total cache size in bytes, invisibly.
Also prints a human-readable size message.
}
\description{
Returns the total size of all files in the cache directory.
Use \code{options(maup.use_cache = TRUE)} to enable persistent caching, or
\code{options(maup.cache_dir = '/path/to/dir')} to specify a custom location.
If caching is not enabled, a temporary directory is used.
}
\examples{
maup_cache_size()
}
