% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maup.R
\name{maup_tract_map}
\alias{maup_tract_map}
\title{Download MAUP county-level maps}
\usage{
maup_tract_map(state, year = 2020, refresh = FALSE)
}
\arguments{
\item{state}{A state name, abbreviation, or FIPS code.}

\item{year}{The redistricting cycle year.}

\item{refresh}{If \code{TRUE}, ignore the cache and download again.}
}
\value{
A named list of \link[redist:redist_map]{redist_map} objects, one per
county.
}
\description{
Downloads a named list of tract-level redistricting maps from the MAUP
Project's Harvard Dataverse, one \link[redist:redist_map]{redist_map} per
county.
}
\examples{
\dontshow{if (Sys.getenv('DATAVERSE_KEY') != '') withAutoprint(\{ # examplesIf}
maup_tract_map('NJ', year = 2020)
\dontshow{\}) # examplesIf}
}
