% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maup.R
\name{maup_tract_plans}
\alias{maup_tract_plans}
\title{Download MAUP county-level plans}
\usage{
maup_tract_plans(state, year = 2020, refresh = FALSE)
}
\arguments{
\item{state}{A state name, abbreviation, or FIPS code.}

\item{year}{The redistricting cycle year.}

\item{refresh}{If \code{TRUE}, ignore the cache and download again.}
}
\value{
A named list of \link[redist:redist_plans]{redist_plans} objects, named
by county FIPS code.
}
\description{
Downloads tract-level redistricting simulation plans from the MAUP
Project's Harvard Dataverse. Each county's plans are downloaded separately
and returned as a named list.
}
\examples{
\dontshow{if (Sys.getenv('DATAVERSE_KEY') != '') withAutoprint(\{ # examplesIf}
maup_tract_plans('NJ', year = 2020)
\dontshow{\}) # examplesIf}
}
