% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rezone}
\alias{rezone}
\title{Rezone existing data}
\usage{
rezone(x, mori, tol = 1/100, maxIter = 10000, print_out = TRUE)
}
\arguments{
\item{x}{an object of class \code{Raster}.}

\item{rc_mat}{a reclassification matrix to passed \link{Reclassify}.}

\item{ras}{logical indicating whether output should be in raster or vector
format.  The default is \code{FALSE}.}
}
\value{
\code{algo3} returns either a \code{Raster} or
\code{SpatialPolygons} class object representing a set of randomly generated
zones.
}
\description{
\code{rezone} randomly rezones an existing zoning system, allowing for
user-defined constraints.  Zones can be output in either raster or vector
format.
}
\details{
Implements the algorithm outlined by Openshaw (1977).
}
\examples{
algo3(x = c(50, 50))
}
\references{
Openshaw, Stan.  1977.  "Algorithm 3: a procedure to
generate pseudo-random aggregations of N zones into M regions, where M is
less than N."  Environment and Planning A 9: 1423-1428.
}

